Added a `BigDecimal(normalize: BigDecimal)` API Initializes a normalized copy of a `BigDecimal`, stripping trailing zeros when the exponent is negative.
Useful when you end up with a BigDecimal filled
with insignificant zeros.

Example occurrence
------------------

Context: Unit of Measure conversions & dimensions
A conversion chain of
```
inch -> feet -> miles -> kilometers -> picometers
-> astronomical units -> inch
```
works like magic using BigDecimal, but the end
result when yielded as a string, contains the
zeros accumulated along the journey.

This contribution aims to provide a solution for
such cases.